Factory Added

+++ app/src/main/java/com/alamin/di/MessageService.kt
@@ -3,7 +3,7 @@ package com.alamin.di
 import android.content.ContentValues
 import android.util.Log
 
-class MessageService(private val retryCount: Int) : NotificationService {
+class MessageService(private val retryCount: Double) : NotificationService {
     override fun send(to: String, form: String, body: String) {
         Log.d(ContentValues.TAG, "User Message Sent With Retry $retryCount")
     }
